[DON-1968] Refactor moved CalendarDayCellTestTag from internal package Moved copyright notice to top of file
